rename folder. remove getType method
