+++ packages/rocket-webhook-core/src/response/responses/webhook-empty-response.ts
